+++ R/prune_fam.R
@@ -5,21 +5,26 @@
 #'
 #' @param fam The pedigree data.frame, in plink FAM format.
 #' Only columns `id`, `pat`, and `mat` are required.
-#' Founders must be present, and their `pat` and `mat` values must be 0 (missing).
-#' Non-founders must have both their parents be non-0.
+#' `id` must be unique and non-missing.
+#' Founders must be present, and their `pat` and `mat` values must be missing (see below).
+#' Non-founders must have both their parents be non-missing.
 #' Parents must appear earlier than their children in the table.
 #' @param ids The list of individuals of interest, whose ancestors we want to keep.
 #' All must be present in `fam$id`.
+#' @param missing_vals The list of ID values treated as missing.
+#' `NA` is always treated as missing.
+#' By default, the empty string ('') and zero (0) are also treated as missing (remove values from here if this is a problem).
 #' 
 #' @return The filtered FAM table with non-ancestors of `ids` excluded.
+#' IDs that are `NA`-equivalent (see `missing_vals`) will be mapped to `NA`.
 #'
 #' @examples
 #' # construct a family with three founders, but one "bob" has no descendants
 #' library(tibble)
 #' fam <- tibble(
 #'     id  = c('mom', 'dad', 'bob', 'child'),
-#'     pat = c(    0,     0,     0,   'dad'),
-#'     mat = c(    0,     0,     0,   'mom')
+#'     pat = c(   NA,    NA,    NA,   'dad'),
+#'     mat = c(   NA,    NA,    NA,   'mom')
 #' )
 #' # only want 'child' and its ancestors
 #' ids <- 'child'
@@ -28,7 +33,7 @@
 #' fam2
 #' 
 #' @export
-prune_fam <- function( fam, ids ) {
+prune_fam <- function( fam, ids, missing_vals = c('', 0) ) {
     if ( missing( fam ) )
         stop( '`fam` is required!' )
     if ( missing( ids ) )
@@ -44,6 +49,15 @@ prune_fam <- function( fam, ids ) {
     if ( is.null( fam$mat ) )
         stop( '`fam$mat` is required!' )
     
+    # normalize missing values for individuals (id, pat, mat)
+    # the only obvious value that is missing no matter what we chose is NA
+    # these others are treated as NA-equivalent
+    for ( missing_val in missing_vals ) {
+        fam$id[ fam$id == missing_val ] <- NA
+        fam$pat[ fam$pat == missing_val ] <- NA
+        fam$mat[ fam$mat == missing_val ] <- NA
+    }
+    
     # in case input is weird, this will behave
     ids <- unique( ids )
     # ids should be in FAM!
@@ -57,12 +71,12 @@ prune_fam <- function( fam, ids ) {
         # collect their parents, make unique (siblings share some)
         parents <- unique( c( fam$pat[ indexes ], fam$mat[ indexes ] ) )
         # and never include people we've already analyzed (relevant for complex pedigrees with individiuals mating across generations)
-        # also exclude missing parents (0)
-        parents <- setdiff( parents, c(ids_keep, 0) )
+        # also exclude missing parents (NA)
+        parents <- setdiff( parents, c(ids_keep, NA) )
         # add to IDs to keep (since we already subtracted shared people, simple concatenation keeps list unique)
         ids_keep <- c( ids_keep, parents )
         # now look for their parents!
-        # NOTE: eventually we hit founders, whose parents are all "0", which we've removed, so list will be empty
+        # NOTE: eventually we hit founders, whose parents are all missing, which we've removed, so list will be empty
         ids <- parents
     }
     # by construction, ids_keep is non-redundant and has no missing IDs